requirements.txt
pyaudio
websockets
openai
requests
pydub
python-dotenv

# List your actual Python dependencies here
speechrecognition
ffmpeg-python